+++ src/components/messenger/list/conversation-item/conversation-item.scss
@@ -1,0 +1,144 @@
+@use '~@zero-tech/zui/styles/theme' as theme;
+@import '../../../../variables';
+
+.conversation-item {
+  display: flex;
+  align-content: stretch;
+  gap: 16px;
+  align-items: center;
+  cursor: pointer;
+  height: 56px;
+  margin-bottom: 8px;
+  padding: 0 8px;
+
+  font-size: $font-size-small;
+  line-height: 17px;
+
+  color: theme.$color-greyscale-12;
+
+  &:hover,
+  &:focus,
+  &[is-active='true'] {
+    background-color: theme.$color-primary-3;
+    border-radius: 8px;
+    outline: none;
+  }
+
+  & > * {
+    flex-shrink: 0;
+  }
+
+  &__summary {
+    flex-shrink: 1;
+    flex-grow: 1;
+    overflow: hidden;
+  }
+
+  &__header {
+    display: flex;
+    align-items: baseline;
+    width: 100%;
+    margin-bottom: 7px;
+    overflow: hidden;
+  }
+
+  &__name {
+    overflow: hidden;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    flex-grow: 1;
+    font-size: $font-size-medium;
+
+    &[is-unread='true'] {
+      font-weight: 700;
+    }
+  }
+
+  &__timestamp {
+    color: theme.$color-greyscale-11;
+    font-weight: 400;
+    font-size: 10px;
+    line-height: 12px;
+    white-space: nowrap;
+  }
+
+  &__message {
+    color: theme.$color-greyscale-11;
+    font-weight: 400;
+    font-size: $font-size-small;
+    line-height: 15px;
+    overflow: hidden;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    padding-right: 1px;
+
+    &[is-unread='true'] {
+      font-weight: 700;
+      color: theme.$color-greyscale-12;
+
+      div {
+        font-weight: 700;
+        color: theme.$color-greyscale-12;
+      }
+    }
+
+    div {
+      overflow: hidden;
+      white-space: nowrap;
+      text-overflow: ellipsis;
+    }
+
+    a {
+      pointer-events: none;
+      /* Disable pointer events for clicking urls */
+    }
+  }
+
+  &__unread-count {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-grow: 0;
+    flex-shrink: 0;
+    background-color: theme.$color-primary-8;
+    color: theme.$color-primary-transparency-12;
+    border-radius: 50%;
+    width: 16px;
+    height: 16px;
+    text-align: center;
+    font-size: 8px;
+    font-weight: 700;
+    line-height: 11px;
+    margin-left: 3px;
+  }
+
+  &__content {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+  }
+
+  &__group-icon {
+    position: relative;
+    width: 40px;
+    height: 40px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    background: linear-gradient(223deg, #421349 14.62%, #0d0416 36.73%, #0b060e 59.58%, #2b0659 85.38%);
+    border-radius: 9999px;
+
+    & > * {
+      margin: auto;
+    }
+  }
+
+  &__group-status {
+    position: absolute;
+    bottom: -0.125rem;
+    right: -0.125rem;
+    border: 0.125rem solid var(--color-primary-1);
+  }
+}
