+++ src/components/messenger/list/conversation-item/conversation-item.test.tsx
@@ -3,8 +3,8 @@ import { shallow } from 'enzyme';
 
 import { ConversationItem, Properties } from './conversation-item';
 import moment from 'moment';
-import { ContentHighlighter } from '../../content-highlighter';
-import { MediaType, MessageSendStatus } from '../../../store/messages';
+import { ContentHighlighter } from '../../../content-highlighter';
+import { MediaType, MessageSendStatus } from '../../../../store/messages';
 
 describe('ConversationItem', () => {
   const subject = (props: Partial<Properties>) => {
