+++ src/components/messenger/list/conversation-item/index.ts
@@ -1,0 +1,1 @@
+export * from './conversation-item';
