+++ src/components/messenger/list/conversation-list-panel/conversation-list-panel.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 
 import { ConversationListPanel, Properties } from './conversation-list-panel';
-import { Channel } from '../../../store/channels';
+import { Channel } from '../../../../store/channels';
 
 describe('ConversationListPanel', () => {
   const subject = (props: Partial<Properties>) => {
