+++ src/components/messenger/list/conversation-list-panel/conversation-list-panel.tsx
@@ -1,19 +1,20 @@
 import * as React from 'react';
 
-import { otherMembersToString } from '../../../platform-apps/channels/util';
-import { Channel } from '../../../store/channels';
+import { otherMembersToString } from '../../../../platform-apps/channels/util';
+import { Channel } from '../../../../store/channels';
 import { IconPlus, IconUserPlus1 } from '@zero-tech/zui/icons';
-import { IconButton } from '../../icon-button';
-import { ConversationItem } from './conversation-item';
-import { InviteDialogContainer } from '../../invite-dialog/container';
+import { IconButton } from '../../../icon-button';
+import { ConversationItem } from '../conversation-item';
+import { InviteDialogContainer } from '../../../invite-dialog/container';
 import { Button, Input, Modal } from '@zero-tech/zui/components';
-import { Item, Option } from '../lib/types';
-import { UserSearchResults } from './user-search-results';
-import { itemToOption } from '../lib/utils';
-import { ScrollbarContainer } from '../../scrollbar-container';
+import { Item, Option } from '../../lib/types';
+import { UserSearchResults } from '../user-search-results/user-search-results';
+import { itemToOption } from '../../lib/utils';
+import { ScrollbarContainer } from '../../../scrollbar-container';
 import escapeRegExp from 'lodash/escapeRegExp';
 
-import { bemClassName } from '../../../lib/bem';
+import { bemClassName } from '../../../../lib/bem';
+import './conversation-list-panel.scss';
 
 const cn = bemClassName('messages-list');
