+++ src/components/messenger/list/create-conversation-panel/create-conversation-panel.test.tsx
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 
 import CreateConversationPanel, { Properties } from './create-conversation-panel';
 
-jest.mock('../autocomplete-members');
+jest.mock('../../autocomplete-members');
 
 describe('CreateConversationPanel', () => {
   const subject = (props: Partial<Properties>) => {
