+++ src/components/messenger/list/group-details-panel/group-details-panel.scss
@@ -1,0 +1,78 @@
+@use '~@zero-tech/zui/styles/theme' as theme;
+@import '../../../../variables';
+@import '../../../../layout';
+
+$side-padding: 16px;
+
+.group-details-panel {
+  & > * {
+    padding-right: $side-padding;
+    padding-left: $side-padding;
+  }
+
+  &__details-content {
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+    margin-top: 8px; // 8px + PanelHeader padding-bottom: 16px for 24px gap.
+  }
+
+  &__selected-count {
+    font-style: normal;
+    font-weight: 400;
+    font-size: $font-size-large;
+    line-height: 24px;
+    padding-top: 10px;
+    color: theme.$color-greyscale-11;
+  }
+
+  &__selected-number {
+    color: theme.$color-greyscale-12;
+  }
+
+  &__create {
+    width: 100%;
+
+    div {
+      display: flex;
+      align-items: center;
+      gap: 8px;
+    }
+  }
+
+  &__field-info-container {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  &__field-info {
+    display: flex;
+    align-items: baseline;
+    justify-content: space-between;
+    width: 100%;
+    line-height: 24px;
+    margin-bottom: 12px;
+    margin-top: 12px;
+
+    &:first-child {
+      margin-top: 0px;
+    }
+  }
+
+  &__label {
+    color: theme.$color-greyscale-12;
+    font-weight: 400;
+    font-size: 16px;
+  }
+
+  &__optional {
+    color: theme.$color-greyscale-11;
+    font-family: 'SF Pro Text';
+    font-size: $font-size-small;
+  }
+
+  .group-details-panel__image-upload {
+    font-size: $font-size-medium;
+  }
+}
