+++ src/components/messenger/list/group-details-panel/index.ts
@@ -1,0 +1,1 @@
+export * from './group-details-panel';
