+++ src/components/messenger/list/panel-header/index.ts
@@ -1,0 +1,1 @@
+export * from './panel-header';
