+++ src/components/messenger/list/selected-user-tag/index.ts
@@ -1,0 +1,1 @@
+export * from './selected-user-tag';
