+++ src/components/messenger/list/selected-user-tag/selected-user-tag.scss
@@ -1,0 +1,37 @@
+@use '~@zero-tech/zui/styles/theme' as theme;
+@import '../../../../variables';
+
+.selected-user-tag {
+  &__selected-option {
+    display: inline-block;
+    vertical-align: top;
+  }
+
+  &__selected-tag {
+    color: theme.$color-greyscale-11;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    padding: 4px 8px;
+    gap: 4px;
+    background-color: theme.$color-greyscale-3;
+    border-radius: 8px;
+    border: 1px solid theme.$color-greyscale-5;
+    margin-top: 10px;
+    margin-right: 10px;
+
+    font-weight: 400;
+    font-size: $font-size-small;
+    line-height: 15px;
+  }
+
+  &__user-remove {
+    cursor: pointer;
+    background: none;
+    border: none;
+    display: inline-block;
+    padding: 0px;
+    margin: 0px;
+  }
+}
