+++ src/components/messenger/list/start-group-panel/index.ts
@@ -1,0 +1,1 @@
+export * from './start-group-panel';
