+++ src/components/messenger/list/user-search-results/index.ts
@@ -1,0 +1,1 @@
+export * from './user-search-results';
